[ISSUE #8208]♻️Extract timer, POP, and local service boundaries
