+++ rocketmq-store-local/Cargo.toml
@@ -37,6 +37,7 @@ observability = [
 [dependencies]
 bytes.workspace = true
 cheetah-string.workspace = true
+dashmap.workspace = true
 memmap2 = "0.9.10"
 page_size = "0.6.0"
 parking_lot.workspace = true
