+++ rocketmq-store-local/src/lib.rs
@@ -16,9 +16,15 @@ pub mod base;
 pub mod commit_log;
 pub mod config;
 pub mod consume_queue;
+pub mod filter;
 pub mod flush;
 pub mod ha;
+pub mod hook;
 pub mod index;
 pub mod mapped_file;
+pub mod pop;
+pub mod services;
+pub mod stats;
+pub mod timer;
 pub mod transfer;
 pub mod utils;
