+++ rocketmq-store-local/src/services.rs
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use rocketmq_store_local::timer::slot::Slot;
+pub mod cold_data_check;
