+++ rocketmq-store/src/filter.rs
@@ -12,46 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::collections::HashMap;
-use std::sync::Arc;
-
-use cheetah_string::CheetahString;
-
-use crate::consume_queue::cq_ext_unit::CqExtUnit;
-
-/// A trait for filtering messages.
-pub trait MessageFilter: Send + Sync {
-    /// Checks if the message is matched by the consume queue.
-    ///
-    /// # Arguments
-    ///
-    /// * `tags_code` - An optional tag code.
-    /// * `cq_ext_unit` - An optional reference to a `CqExtUnit`.
-    ///
-    /// # Returns
-    ///
-    /// * `true` if the message is matched, `false` otherwise.
-    fn is_matched_by_consume_queue(&self, tags_code: Option<i64>, cq_ext_unit: Option<&CqExtUnit>) -> bool;
-
-    /// Checks if the message is matched by the commit log.
-    ///
-    /// # Arguments
-    ///
-    /// * `msg_buffer` - An optional mutable reference to a message buffer.
-    /// * `properties` - An optional reference to a map of properties.
-    ///
-    /// # Returns
-    ///
-    /// * `true` if the message is matched, `false` otherwise.
-    fn is_matched_by_commit_log(
-        &self,
-        msg_buffer: Option<&[u8]>,
-        properties: Option<&HashMap<CheetahString, CheetahString>>,
-    ) -> bool;
-}
-
-/// Type alias for an atomically reference-counted message filter.
-///
-/// This is the standard way to share a [`MessageFilter`] implementation across multiple
-/// threads or components. The `Arc` provides thread-safe reference counting.
-pub type ArcMessageFilter = Arc<dyn MessageFilter>;
+pub use rocketmq_store_local::filter::ArcMessageFilter;
+pub use rocketmq_store_local::filter::MessageFilter;
