+++ rocketmq-store/src/log_file/cold_data_check_service.rs
@@ -12,41 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use cheetah_string::CheetahString;
-
-/// Service for checking if message data is in cold storage area
-#[derive(Default)]
-pub struct ColdDataCheckService;
-
-impl ColdDataCheckService {
-    /// Check if the data at the given offset is in page cache
-    pub fn is_data_in_page_cache(&self) -> bool {
-        true
-    }
-
-    /// Check if the message at the given queue offset is in cold data area
-    ///
-    /// Cold data is data that has not been accessed for a long time and may be
-    /// stored on slower storage or has been paged out from memory.
-    ///
-    /// # Arguments
-    /// * `consumer_group` - The consumer group name
-    /// * `topic` - The topic name
-    /// * `queue_id` - The queue ID
-    /// * `queue_offset` - The queue offset to check
-    ///
-    /// # Returns
-    /// `true` if the message is in cold data area, `false` otherwise
-    pub fn is_msg_in_cold_area(
-        &self,
-        _consumer_group: &CheetahString,
-        _topic: &CheetahString,
-        _queue_id: i32,
-        _queue_offset: i64,
-    ) -> bool {
-        // TODO: Implement actual cold data detection logic
-        // This should check if the message's physical offset is in cold storage area
-        // based on configuration like cold data threshold time or physical offset threshold
-        false
-    }
-}
+pub use rocketmq_store_local::services::cold_data_check::ColdDataCheckService;
